requirements.txt
# Core dependencies
alpaca-trade-api>=3.0.0
alpaca-py>=0.8.0

# Data handling
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# API and networking
requests>=2.28.0
websocket-client>=1.5.0

# Utilities
tzlocal>=4.3
schedule>=1.2.0
click>=8.1.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0

# Optional: Notifications
discord-webhook>=1.1.0
slack-sdk>=3.21.0

# Optional: Data visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Type hints
types-requests>=2.31.0
types-PyYAML>=6.0.0
